moved functions to different namespaces + modified docs
