+++ src/functions/http/ping.js
