+++ src/functions/system/pause.js
