+++ src/functions/system/setWindowTitle.js
