+++ src/functions/system/softShutdown.js
